Unnecessary unboxing

+++ dockstore-webservice/src/main/java/io/dockstore/webservice/jdbi/EventDAO.java
@@ -52,7 +52,7 @@ public class EventDAO extends AbstractDAO<Event> {
     public long countAllEventsForOrganization(long organizationId) {
         final Query query = namedQuery("io.dockstore.webservice.core.Event.countAllForOrganization")
                 .setParameter("organizationId", organizationId);
-        return ((Long)query.getSingleResult()).longValue();
+        return (Long) query.getSingleResult();
     }
 
     public List<Event> findEventsByEntryIDs(Set<Long> entryIds, Integer offset, int limit) {
